# backend/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FeedTrack API",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://feedtrack-backend.vercel.app",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/branches/areas": {
      "get": {
        "summary": "Get unique areas from branches",
        "description": "Get unique areas from the 'branch' table",
        "responses": {
          "200": {
            "description": "List of unique areas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "Kanton 10",
                      "Hercegovačko-neretvanski kanton",
                      "Bosansko-podrinjski kanton",
                      "Srednjobosanski kanton",
                      "Kanton Sarajevo",
                      "Tuzlanski kanton",
                      "Zeničko-dobojski kanton",
                      "Republika Srpska"
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching unique areas:",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/branches/by-area/{area}": {
      "get": {
        "summary": "Get branches by area",
        "description": "Get branches by area from the 'branch' table",
        "parameters": [
          {
            "name": "area",
            "in": "path",
            "description": "Area of the branch",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branches in the specified area",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Branch"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching branches by area:",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignID}/questionsPerPage": {
      "get": {
        "summary": "Get questions per page for a campaign",
        "description": "Get the number of questions per page for a specific campaign",
        "parameters": [
          {
            "name": "campaignID",
            "in": "path",
            "description": "ID of the campaign",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Questions per page for the specified campaign",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "questionsPerPage": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching questions per page:",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/{campaignID}/campaignQuestions": {
      "get": {
        "summary": "Get all questions for a campaign",
        "description": "Get all questions for a specific campaign",
        "parameters": [
          {
            "name": "campaignID",
            "in": "path",
            "description": "ID of the campaign",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of questions for the specified campaign",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example":"3"
                    },
                    "name": {
                      "type": "string",
                      "example":"Prvo pitanje"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Campaign questions not found",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Campaign questions not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching campaign questions:",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/welcomeData": {
        "get": {
          "summary": "Get welcome message data",
          "responses": {
            "200": {
              "description": "Successful response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message":{
                        "type":"string",
                        "example":"Hello"
                      }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error:",
              "content": {
                "application/json": {
                  "schema": {
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Upload welcome message data",
          "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/WelcomeMessage"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "File and message uploaded successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "File and message uploaded successfully"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request",
              "content": {
                "application/json": {
                  "schema": {
                    "oneOf": [
                      {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "enum": ["Message is required", "No file uploaded", "Content-Type must be multipart/form-data"]
                          }
                        }
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "example": "Message is required"
                        }
                      }
                    ]
                  }
                }
              }
            },          
            "500": {
              "description": "Internal Server Error:",
              "content": {
                "application/json": {
                  "schema": {
                    "properties": {
                      "error": {
                        "type": "string",
                        "example": "Internal server error"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Branch": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64",
            "example": 19
          },
          "location": {
            "type": "string",
            "example": "Zenica"
          },
          "managerID": {
            "type": "string",
            "example": 1
          },
          "rating": {
            "type": "integer",
            "example": 4
          },
          "area": {
            "type": "string",
            "example": "Zeničko-dobojski kanton"
          },
          "campaignID": {
            "type": "string",
            "example": 2
          }
        }
      },
...
      "WelcomeMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Hello"
          }
        }
      }
    }
  }
}